requirements.txt
annotated-types==0.6.0
anyio==4.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
deprecation==2.1.0
feedparser==6.0.10
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
install==1.3.5
packaging==23.2
postgrest==0.13.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.0
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.1.0
supafunc==0.3.1
typing_extensions==4.8.0
urllib3==2.1.0
websockets==10.4
